Разбиение файла на модули
